refactor : 리뷰 반영
- dto 형식 통일
- 불필요 dto 삭제
- post, patch response message 추가

+++ src/main/java/com/coderder/colorMeeting/dto/request/TeamTimeDto.java
@@ -5,6 +5,6 @@ import lombok.Getter;
 @Getter
 public class TeamTimeDto {
     private Long teamId;
-    private Long spendingMinute;
+    private int spendingMinute;
 
 }
